関係者のadd, removeをcomment方式に変更

+++ app/controllers/key_results_controller.rb
@@ -19,7 +19,7 @@ class KeyResultsController < ApplicationController
     ActiveRecord::Base.transaction do
       @key_result = KeyResult.find(params[:id])
       @key_result.update!(key_result_update_params)
-      update_concerned_people if params[:key_result][:concerned_people]
+      update_concerned_people if params[:key_result][:concerned_person]
       update_comment if params[:key_result][:comment]
     end
     render action: :create, status: :ok
@@ -39,15 +39,11 @@ class KeyResultsController < ApplicationController
   private
 
   def update_concerned_people
-    current_people = @key_result.concerned_people.pluck(:user_id)
-    new_people = params[:key_result][:concerned_people]
-    add_list = new_people - current_people
-    remove_list = current_people - new_people
-    add_list.each do |id|
-      @key_result.concerned_people.create!(user_id: id, role: 0)
-    end
-    remove_list.each do |id|
-      person = @key_result.concerned_people.find_by(user_id: id)
+    concerned_person_data = params[:key_result][:concerned_person]
+    if concerned_person_data['behavior'] == 'add'
+      @key_result.concerned_people.create!(user_id: concerned_person_data['data'], role: 0)
+    elsif concerned_person_data['behavior'] == 'remove'
+      person = @key_result.concerned_people.find_by(user_id: concerned_person_data['data'])
       person.destroy!
     end
   end
